% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar.R
\name{has_conn}
\alias{has_conn}
\title{Verifica conexão}
\usage{
has_conn()
}
\value{
TRUE se acessa o site da Receita, FALSE caso contrário.
}
\description{
Verifica se é possível acessar a página de busca de CNPJ da Receita.
}

